Move TestingDelegate to remote-batch-store-fake.js
