+++ js/noms/src/remote-batch-store-test.js
@@ -6,12 +6,12 @@
 
 import {suite, test} from 'mocha';
 import {assert} from 'chai';
-import {makeTestingRemoteBatchStore} from './remote-batch-store.js';
+import makeRemoteBatchStoreFake from './remote-batch-store-fake.js';
 import {encodeValue} from './codec.js';
 
 suite('BatchStore', () => {
   test('get after schedulePut works immediately', async () => {
-    const bs = makeTestingRemoteBatchStore();
+    const bs = makeRemoteBatchStoreFake();
     const input = 'abc';
 
     const c = encodeValue(input);
@@ -23,7 +23,7 @@ suite('BatchStore', () => {
   });
 
   test('get after schedulePut works after flush', async () => {
-    const bs = makeTestingRemoteBatchStore();
+    const bs = makeRemoteBatchStoreFake();
     const input = 'abc';
 
     const c = encodeValue(input);
